[🚚move]: useObserver hook directory 위치 변경
