+++ src/app/component/page.tsx
@@ -21,7 +21,7 @@ import {
 import { MainContainer } from "@/components/Pages";
 import useChipStore from "@/store/component/useChipStore";
 import useContextMenuStore from "@/store/common/useContextMenuStore";
-import { useObserver } from "@/hooks/api/common/useObserver";
+import { useObserver } from "@/hooks/common/useObserver";
 import useComponentListInfiniteQuery from "@/hooks/api/component/useComponentListInfiniteQuery";
 import {
   COMPONENT_SORT_CONDITION,
